Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-    int radius;
-    scanf("%d",&radius);
-    printf("Area:%d",Area=3.14*radius*radius)
+    float radius;
+    scanf("%f",&radius);
+    printf("Area:%.2f",Area=3.14*radius*radius)
     return 0;
 }
